#16 refactor detailspage
