+++ src/Routes/Details.tsx
@@ -1,72 +1,25 @@
 import React, { useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
+import { useParams,useNavigate } from 'react-router-dom';
 import './Details.css';
 import { useSelector, useDispatch } from 'react-redux';
 import { addFavoriteMovie, removeFavoriteMovie } from '../store/actions';
-import { useNavigate } from 'react-router-dom';
-
-interface CrewMember {
-  job: string;
-  department: string;
-  name: string;
-}
-
-interface CastMember {
-  id: number;
-  name: string;
-}
-
-interface Movie {
-  id: number;
-  original_title: string;
-  poster_path: string;
-  genres: { id: number; name: string }[];
-  release_date: string;
-  overview: string;
-  status: string;
-  vote_average: number;
-  vote_count: number;
-  budget: number;
-}
+import Loading from '../Components/Loading';
+import detailsFetch from '../Services/detailsFetch';
+import { Movie, CrewMember, CastMember} from "../Services/detailsTypes"
 
 const Details = () => {
-  const { showId } = useParams<{ showId: string }>();
-  const [movie, setMovie] = useState<Movie | null>(null);
+  const { showId } = useParams<{ showId: string |undefined }>();
+  const [movie, setMovie] = useState<Movie| null>(null);
   const [credits, setCredits] = useState<any | null>(null);
-  const navigate = useNavigate();
   const [director, setDirector] = useState<CrewMember | null>(null);
   const [mainCast, setMainCast] = useState<CastMember[] | null>(null);
-
-  const dispatch = useDispatch();
   const favoriteMovies = useSelector((state: any) => state.favoriteMovies);
+  const navigate = useNavigate();
+  const dispatch = useDispatch();
+  const isLoading = detailsFetch({showId, setMovie, setMainCast, setDirector, setCredits, dispatch})
 
-  const myapikey = `6ef10486c5df46ca61884c8b042d53bd`
-  const castUrl = `https://api.themoviedb.org/3/movie/${showId}/credits?api_key=${myapikey}`;
-  const showUrl = `https://api.themoviedb.org/3/movie/${showId}?api_key=${myapikey}`;
-
-  useEffect(() => {
-    const fetchMovieDetails = async () => {
-      try {
-        const movieResponse = await fetch(showUrl);
-        const movieData = await movieResponse.json();
-        setMovie(movieData);
-
-        const castResponse = await fetch(castUrl);
-        const castData = await castResponse.json();
-        setCredits(castData);
-        setDirector(castData.crew.find((person: CrewMember) => person.job === 'Director' && person.department === 'Directing'));
-        setMainCast(castData.cast.slice(0, 3));
-      } catch (error) {
-        console.error('Error fetching movie details:', error);
-      }
-    };
-
-    fetchMovieDetails();
-  }, [showId, dispatch]);
-
-  if (!movie) {
-    return <p>Loading...</p>;
-  }
+  if (!movie) {return (<Loading/>);}
+  if (isLoading) {return (<Loading/>);}
 
   const isMovieFavorite = favoriteMovies.some((favMovie: Movie) => favMovie.id === movie.id);
