+++ src/Services/tmdbApiServices.tsx
@@ -1,8 +1,10 @@
 
 
-export const apiKey : string = `6ef10486c5df46ca61884c8b042d53bd`;
-export const baseUrl : string = `https://api.themoviedb.org/3/`;
+export const apiKey: string = `6ef10486c5df46ca61884c8b042d53bd`;
+export const baseUrl: string = `https://api.themoviedb.org/3/`;
 
+//spesifik urller burada
 export const genreListUrl = `${baseUrl}genre/movie/list?api_key=${apiKey}`;
+
